feat: add sweetalert

+++ src/app/(auth)/issues/[owner]/[repo]/[number]/components/DeleteIssueButton.tsx
@@ -8,6 +8,19 @@ import axios from "axios";
 import { IssueNumberPageParams } from "@/types";
 
 import { FaRegTrashAlt } from "react-icons/fa";
+import Swal from "sweetalert2";
+
+const Toast = Swal.mixin({
+    toast: true,
+    position: "top-end",
+    showConfirmButton: false,
+    timer: 1500,
+    timerProgressBar: true,
+    didOpen: (toast) => {
+        toast.onmouseenter = Swal.stopTimer;
+        toast.onmouseleave = Swal.resumeTimer;
+    },
+});
 
 const DeleteIssueButton = ({ params }: IssueNumberPageParams) => {
     const { owner, repo, number } = params;
@@ -19,20 +32,58 @@ const DeleteIssueButton = ({ params }: IssueNumberPageParams) => {
         e: React.MouseEvent<HTMLButtonElement, MouseEvent>
     ) => {
         e.preventDefault();
-        try {
-            const { status } = await axios.delete(
-                `/api/github/issues/${owner}/${repo}/${number}`,
-                {
-                    headers: { Authorization: session?.accessToken },
-                }
-            );
 
-            if (status === 200) {
-                alert("delete successfully");
-                router.push(`/issues/${owner}/${repo}`);
+        const result = await Swal.fire({
+            title: "Are you sure?",
+            text: "You won't be able to revert this!",
+            icon: "warning",
+            showConfirmButton: false,
+            showCancelButton: true,
+            showDenyButton: true,
+            denyButtonText: "Delete",
+        });
+
+        if (result.isDenied) {
+            Swal.fire({
+                icon: "info",
+                title: "Deleting Issue",
+                text: "Please wait...",
+                showConfirmButton: false,
+                allowOutsideClick: false,
+                didOpen: () => {
+                    Swal.showLoading();
+                },
+            });
+
+            try {
+                const { status } = await axios.delete(
+                    `/api/github/issues/${owner}/${repo}/${number}`,
+                    {
+                        headers: { Authorization: session?.accessToken },
+                    }
+                );
+
+                Swal.close();
+
+                if (status === 200) {
+                    Toast.fire({
+                        icon: "success",
+                        title: "Delete successfully",
+                    });
+
+                    router.push(`/issues/${owner}/${repo}`);
+                }
+            } catch (error) {
+                Swal.close();
+
+                Swal.fire({
+                    icon: "error",
+                    title: "Oops...",
+                    text: "Something error",
+                });
+
+                console.log(error.response);
             }
-        } catch (error) {
-            console.log(error.response);
         }
     };
